# Cherry Pick: Speed up shader compilation and avoid WebGL crashes (#8950)