--- a/frontend/javascripts/viewer/controller/viewmodes/arbitrary_controller.tsx
+++ b/frontend/javascripts/viewer/controller/viewmodes/arbitrary_controller.tsx
@@ -328,6 +328,10 @@ class ArbitraryController extends React.PureComponent<Props> {
         },
       ),
       listenToStoreProperty(
+        (state) => state.datasetConfiguration.interpolation,
+        (interpolation) => this.plane.setLinearInterpolationEnabled(interpolation),
+      ),
+      listenToStoreProperty(
         (state) => state.temporaryConfiguration.flightmodeRecording,
         (isRecording) => {
           if (isRecording) {