--- a/unreleased_changes/8950.md
+++ b/unreleased_changes/8950.md
@@ -1,0 +1,2 @@
+### Fixed
+- Fixed WebGL crashes for datasets with many layers and sped up shader compilation time.